refactor: restructure project
Restructuring the project since it is growing in size. Functionality is
now split in cli, core, markdown and defaults modules.
